+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 - Limit the keyboard buffer to 8 item
 - Move the screen convertion from the device to the render thread
+- Huge improvement in the debugger interface
 
 ## [0.4.0]